moved rich edit tag helpers to separate name space
