+++ source/DasBlog.Web.UI/Config/site.config
@@ -177,7 +177,7 @@
   <AllowOpenIdAdmin>false</AllowOpenIdAdmin>
   <BypassSpamOpenIdComment>false</BypassSpamOpenIdComment>
   <TitlePermalinkSpaceReplacement>-</TitlePermalinkSpaceReplacement>
-  <EntryEditControl>TinyMce</EntryEditControl>
+  <EntryEditControl>TextArea</EntryEditControl>
   <AMPPagesEnabled>false</AMPPagesEnabled>
   <RSSEndPointRewrite></RSSEndPointRewrite>
 </SiteConfig>
