+++ source/DasBlog.Web.UI/Startup.cs
@@ -28,6 +28,7 @@ using DasBlog.Core.Services;
 using DasBlog.Core.Services.Interfaces;
 using DasBlog.Web.Common;
 using DasBlog.Web.TagHelpers;
+using DasBlog.Web.TagHelpers.RichEdit;
 
 namespace DasBlog.Web
 {
