+++ source/DasBlog.Web.UI/TagHelpers/RichEdit/IRichEditBuilder.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Razor.TagHelpers;
 using Newtonsoft.Json.Bson;
 
-namespace DasBlog.Web.TagHelpers
+namespace DasBlog.Web.TagHelpers.RichEdit
 {
 	public interface IRichEditBuilder
 	{
