+++ source/DasBlog.Web.UI/TagHelpers/RichEdit/RichEditTagHelper.cs
@@ -1,22 +1,23 @@
 ﻿using System.Threading.Tasks;
 using Microsoft.AspNetCore.Razor.TagHelpers;
 
-namespace DasBlog.Web.TagHelpers
+namespace DasBlog.Web.TagHelpers.RichEdit
 {
-	public class RichEditScriptsTagHelper : TagHelper
+	public class RichEditTagHelper : TagHelper
 	{
-		public string ControlId { get; set; }
+		public string Id { get; set; }
+		public string Name { get; set; }
 
 		private readonly IRichEditBuilder builder;
 
-		public RichEditScriptsTagHelper(IRichEditBuilder builder)
+		public RichEditTagHelper(IRichEditBuilder builder)
 		{
 			this.builder = builder;
 		}
 		
 		public override void Process(TagHelperContext context, TagHelperOutput output)
 		{
-			builder.ProcessScripts(this, context, output);
+			builder.ProcessControl(this, context, output);
 		}
 		public override Task ProcessAsync(TagHelperContext context, TagHelperOutput output)
 		{
