+++ source/DasBlog.Web.UI/TagHelpers/RichEdit/TinyMceBuilder.cs
@@ -1,13 +1,16 @@
 ﻿using Microsoft.AspNetCore.Razor.TagHelpers;
 
-namespace DasBlog.Web.TagHelpers
+namespace DasBlog.Web.TagHelpers.RichEdit
 {
-	public class NicEditBuilder : IRichEditBuilder
+	public class TinyMceBuilder : IRichEditBuilder
 	{
-		private const string NIC_EDIT_SERVICE_URL = "http://js.nicedit.com/nicEdit-latest.js";
-
+		private const string TINY_MCE_SERVICE_URL = "https://cloud.tinymce.com/stable/tinymce.min.js";
 		private const string INIT_SCRIPT_TEMPLATE = @"
-		<script type=""text/javascript"">area1 = new nicEditor({{fullPanel : true}}).panelInstance('{0}',{{hasPanel : true}});</script>
+		<script>
+		tinymce.init({{
+			selector: '#{0}'
+		}});
+		</script>
 		";
 		
 		public void ProcessControl(RichEditTagHelper tagHelper, TagHelperContext context, TagHelperOutput output)
@@ -24,7 +27,7 @@ namespace DasBlog.Web.TagHelpers
 		{
 			output.TagName = "script";
 			output.TagMode = TagMode.StartTagAndEndTag;
-			output.Attributes.SetAttribute("src", NIC_EDIT_SERVICE_URL);
+			output.Attributes.SetAttribute("src", TINY_MCE_SERVICE_URL);
 			string htmlContent = string.Format(INIT_SCRIPT_TEMPLATE, tagHeelper.ControlId);
 			output.PostElement.SetHtmlContent(htmlContent);
 		}
